Dont use dotfile for env
It's debatable, but imo it's clearer not to make env a dotfile.
Typically we expect to see dotfiles in a users home directory, not in a
tool repo.

The deploy scripts are part of a specific install process and won't be
ran regularly by a user. The values in the env file should be secured
and stored appropriately; it's not something that should live in a users
home directory.